backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
openai>=1.40.0
python-dotenv>=1.0.0
python-multipart>=0.0.9
pydantic>=1.10.0
google-generativeai>=0.6.0
sqlalchemy>=2.0.0
google-cloud-speech>=2.26.0
sentry-sdk>=1.40.0
psycopg[binary]>=3.1.0
